+++ src/package/components/input-addon.vue
@@ -1,0 +1,15 @@
+<template>
+    <component :is="as" class="vf-input-group__addon">
+        <slot />
+    </component>
+</template>
+
+<script setup lang="ts">
+interface Props {
+    as?: string;
+}
+
+withDefaults(defineProps<Props>(), {
+    as: 'span',
+});
+</script>