+++ src/package/themes/default/components/input-group.ts
@@ -1,0 +1,19 @@
+export default {
+    gap: '0',
+    borderRadius: 'var(--vf-radii-md)',
+    addonPadding: '0 var(--vf-controls-padding-x)',
+    addonFontSize: 'var(--vf-typography-font-size)',
+    addonBackgroundColor: 'var(--vf-controls-background-color)',
+    addonOutlinedBackgroundColor: 'var(--vf-bg-color)',
+    addonTextColor: 'var(--vf-secondary-text-color)',
+    addonBorderColor: 'var(--vf-border-color)',
+    disabledOpacity: 'var(--vf-states-disabled-opacity)',
+    small: {
+        addonPadding: '0 var(--vf-sizes-sm-padding-x)',
+        addonFontSize: 'var(--vf-sizes-sm-font-size)',
+    },
+    large: {
+        addonPadding: '0 var(--vf-sizes-lg-padding-x)',
+        addonFontSize: 'var(--vf-sizes-lg-font-size)',
+    },
+};